refactor(api): extract rack support into dedicated crate
Move rack-related clients, firmware helpers, and RMS test support out of the
API crate into a new carbide-rack crate. Also move shared MQTT publishing
helpers into carbide-mqtt-common so rack/BMS publishing can preserve existing
behavior outside the API crate.

This is a behavior-preserving code move in preparation for extracting rack and
power-shelf controllers from the API crate.
