+++ crates/api/src/handlers/rack.rs
@@ -19,6 +19,7 @@ use std::str::FromStr;
 
 use ::rpc::errors::RpcDataConversionError;
 use ::rpc::forge::{self as rpc, HealthReportEntry};
+use carbide_rack::firmware_object::rack_maintenance_access_token_key;
 use carbide_uuid::machine::MachineId;
 use carbide_uuid::power_shelf::PowerShelfId;
 use carbide_uuid::rack::RackId;
@@ -38,7 +39,6 @@ use tonic::{Request, Response, Status};
 use crate::CarbideError;
 use crate::api::{Api, log_request_data, log_request_data_redacted};
 use crate::auth::AuthContext;
-use crate::rack::firmware_object::rack_maintenance_access_token_key;
 
 pub async fn get_rack(
     api: &Api,
