+++ crates/api/src/lib.rs
@@ -57,7 +57,6 @@ mod machine_validation;
 mod measured_boot;
 mod mqtt_state_change_hook;
 mod network_segment;
-mod rack;
 mod redfish;
 mod run;
 mod scout_stream;
