+++ crates/rack/Cargo.toml
@@ -1,0 +1,53 @@
+#
+# SPDX-FileCopyrightText: Copyright (c) 2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
+# SPDX-License-Identifier: Apache-2.0
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+# http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+#
+[package]
+name = "carbide-rack"
+version = "0.0.0"
+edition.workspace = true
+license.workspace = true
+authors.workspace = true
+
+[features]
+default = []
+test-support = []
+
+[dependencies]
+bms-dsx-exchange = { path = "../bms-dsx-exchange", default-features = false }
+carbide-api-db = { path = "../api-db", default-features = false }
+carbide-api-model = { path = "../api-model", default-features = false }
+carbide-health-report = { path = "../health-report", default-features = false }
+carbide-mqtt-common = { path = "../mqtt-common", default-features = false }
+carbide-secrets = { path = "../secrets", default-features = false }
+carbide-uuid = { path = "../uuid", default-features = false }
+mqttea = { path = "../mqttea" }
+state-controller = { path = "../state-controller" }
+
+async-trait = { workspace = true }
+chrono = { workspace = true }
+eyre = { workspace = true }
+librms = { workspace = true }
+mac_address = { workspace = true }
+opentelemetry = { workspace = true }
+serde_json = { workspace = true }
+sqlx = { workspace = true }
+tracing = { workspace = true }
+tokio = { workspace = true }
+tokio-util = { workspace = true }
+tonic = { workspace = true }
+
+[lints]
+workspace = true
