+++ crates/rack/src/firmware_object.rs
